Renamed bundle namespaces according to Symfony best practice - part 3
