+++ doc/Development_Documentation/09_Development_Tools_and_Details/09_Cache/README.md
@@ -21,10 +21,10 @@ for details.
 
 ## Configuring the cache
 
-The `PimcoreBundle` defines a default cache configuration which you can override in your config files:
+The `PimcoreCoreBundle` defines a default cache configuration which you can override in your config files:
 
 ```yaml
-# pimcore/lib/Pimcore/Bundle/PimcoreBundle/Resources/config/pimcore/config.yml
+# pimcore/lib/Pimcore/Bundle/CoreBundle/Resources/config/pimcore/config.yml
 
 pimcore:
     cache:
@@ -71,7 +71,7 @@ is enabled as well.
 
 If you want to use a custom cache pool, ignore the `pools` section (or disable both predefined pools) and set the `pool_service_id`
 entry to the service ID of your custom pool (needs to be defined as service on the service container). There are a couple
-of cache pools predefined in [cache.yml](https://github.com/pimcore/pimcore/blob/master/pimcore/lib/Pimcore/Bundle/PimcoreBundle/Resources/config/cache.yml)
+of cache pools predefined in [cache.yml](https://github.com/pimcore/pimcore/blob/master/pimcore/lib/Pimcore/Bundle/CoreBundle/Resources/config/cache.yml)
 but those (array, filesystem) are mainly used for testing. 
 
 > If all of the predefined cache pools are disabled, the cache will fall back to a filesystem cache which is rather slow.
